change main.player_bet_input from int to float

+++ main.py
@@ -8,8 +8,8 @@ def winCheck():
 	pass
 
 player_name_input = input("Enter your name: ")
-player_bet_input = input("Enter the amount of money you'd like to bet: $")
-player_hand = handClass.Hand(str(player_name_input), int(player_bet_input))
+player_bet_input = float(input("Enter the amount of money you'd like to bet: $"))
+player_hand = handClass.Hand(str(player_name_input), player_bet_input)
 dealer_hand = handClass.Hand('dealer', 0)
 
 def playerHits():
